Fix vi mode indicator
zsh provides a `zle-keymap-select` function that triggers every time the
keymode is changed (from insert to command mode). That function was
declared on `spaceship_vi_mode_enable` but still didn't work.

I think that because of the async code and the variables not
propagating, the function couldn't actually redraw the right thing with
an updated prompt. What I did was to add the function to re-do the
prompt on the `zle-keymap-select`.

That said, it will only work if you use the `spaceship_vi_mode_enable`
function somewhere when initializing your shell, like on `.zshrc`. Also
make sure to put it after other stuff that may redefine the function,
like oh-my-zsh vi-mode plugin.